chore: remove unused setNotes setter
